tests/wpt/tests/close-watcher/user-activation/nn-activate-CloseWatcher.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/common/top-layer.js"></script>
<script src="../resources/helpers.js"></script>

<body>
<script>
const type = "CloseWatcher";

promise_test(async t => {
  const events = [];

  createRecordingCloseWatcher(t, events, "watcher1", type);
  const watcher2 = createRecordingCloseWatcher(t, events, "watcher2", type);

  await maybeTopLayerBless(watcher2);

  await sendCloseRequest();
  await waitForPotentialCloseEvent();
  assert_array_equals(events, ["watcher2 cancel[cancelable=true]", "watcher2 close", "watcher1 cancel[cancelable=true]", "watcher1 close"]);
}, "Create two CloseWatchers without user activation; send user activation");
</script>
